+++ src/main/java/eightbit/Main.java
@@ -1,12 +1,14 @@
+package eightbit;
+
 /**
- * Main driver class.
+ * eightbit.Main driver class.
  */
 public class Main {
 
     private static final String FILEPATH = "data/eightBit.txt";
 
     /**
-     * Main driver method.
+     * eightbit.Main driver method.
      */
     public static void main(String[] args) {
         EightBit eightBit = new EightBit(FILEPATH);
