+++ src/main/java/eightbit/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package eightbit.command;
+
+import eightbit.util.Storage;
+import eightbit.util.TaskList;
+import eightbit.util.Ui;
+
 public class ByeCommand extends Command {
 
     @Override
