+++ src/main/java/eightbit/command/ToDoCommand.java
@@ -1,3 +1,10 @@
+package eightbit.command;
+
+import eightbit.task.ToDo;
+import eightbit.util.Storage;
+import eightbit.util.TaskList;
+import eightbit.util.Ui;
+
 public class ToDoCommand extends Command {
 
     private final ToDo toDo;
