+++ src/main/java/eightbit/util/Parser.java
@@ -1,3 +1,11 @@
+package eightbit.util;
+
+import eightbit.EightBitException;
+import eightbit.command.*;
+import eightbit.task.Deadline;
+import eightbit.task.Event;
+import eightbit.task.ToDo;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeParseException;
